+++ src/App.js
@@ -1,9 +1,8 @@
-import RaceController from './service/RaceController';
+import RaceController from './service/RaceService';
 
 class App {
   async run() {
     const service = new RaceController();
-
     service.start();
   }
 }
